+++ src/beamngpy/__init__.py
@@ -3,6 +3,9 @@ BeamNGPy API module.
 
 .. moduleauthor:: Marc Müller <mmueller@beamng.gmbh>
 """
+
+__version__ = '1.16'
+
 from . import beamngcommon, sensors
 
 from .scenario import *