+++ CHANGELOG.rst
@@ -2,6 +2,19 @@
 Changelog
 =========
 
+Version 1.16
+============
+- Add option to specify multiple positions/directions for the ``Camera`` sensor
+- Add option to specify additional Lua extensions to load per vehicle
+- Add ``set_lights`` method to vehicle class
+- Add test for setting lights
+- Add test for vehicle bounding box
+- Add ``over_objects`` field to ``Road`` class
+- Add ``teleport`` flag to ``ai_set_script`` of ``Vehicle`` class
+- Fix lack of `__version__`
+- Fix electrics sensor not returning values directly
+- Fix `ai_set_script` teleporting vehicle
+
 Version 1.15
 ============
 - Add option to pass additional Lua extensions to be loaded on startup